Move zipkin http collector related stuff to subpackage
