+++ cmd/collector/app/http_handler.go
@@ -21,7 +21,6 @@
 package app
 
 import (
-	"compress/gzip"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -33,28 +32,24 @@ import (
 	tchanThrift "github.com/uber/tchannel-go/thrift"
 
 	tJaeger "github.com/uber/jaeger/thrift-gen/jaeger"
-	"github.com/uber/jaeger/thrift-gen/zipkincore"
 )
 
 const (
 	formatParam               = "format"
-	unableToReadBodyErrFormat = "Unable to process request body: %v"
+	UnableToReadBodyErrFormat = "Unable to process request body: %v"
 )
 
 // APIHandler handles all HTTP calls to the collector
 type APIHandler struct {
 	jaegerBatchesHandler JaegerBatchesHandler
-	zipkinSpansHandler   ZipkinSpansHandler
 }
 
 // NewAPIHandler returns a new APIHandler
 func NewAPIHandler(
 	jaegerBatchesHandler JaegerBatchesHandler,
-	zipkinSpansHandler ZipkinSpansHandler,
 ) *APIHandler {
 	return &APIHandler{
 		jaegerBatchesHandler: jaegerBatchesHandler,
-		zipkinSpansHandler:   zipkinSpansHandler,
 	}
 }
 
@@ -63,39 +58,11 @@ func (aH *APIHandler) RegisterRoutes(router *mux.Router) {
 	router.HandleFunc("/api/traces", aH.saveSpan).Methods(http.MethodPost)
 }
 
-func (aH *APIHandler) RegisterZipkinRoutes(router *mux.Router) {
-	router.HandleFunc("/api/v1/spans", aH.zipkinHandler).Methods(http.MethodPost)
-}
-
-func (aH *APIHandler) zipkinHandler(w http.ResponseWriter, r *http.Request) {
-	bRead := r.Body
-	if strings.Contains(r.Header.Get("Content-Encoding"), "gzip") {
-		result, err := gzip.NewReader(r.Body)
-		if err != nil {
-			http.Error(w, "Wrong encoding", http.StatusBadRequest)
-		}
-		bRead = result
-	}
-
-	bodyBytes, err := ioutil.ReadAll(bRead)
-	if err != nil {
-		http.Error(w, "Could not read request body", http.StatusBadRequest)
-	}
-
-	if r.Header.Get("Content-Type") == "application/x-thrift" {
-		handleZipkinThrift(aH.zipkinSpansHandler, bodyBytes, w)
-	} else {
-		http.Error(w, "Only Content-Type:application/x-thrift is supported at the moment", http.StatusBadRequest)
-	}
-
-	w.WriteHeader(http.StatusAccepted)
-}
-
 func (aH *APIHandler) saveSpan(w http.ResponseWriter, r *http.Request) {
 	bodyBytes, err := ioutil.ReadAll(r.Body)
 	r.Body.Close()
 	if err != nil {
-		http.Error(w, fmt.Sprintf(unableToReadBodyErrFormat, err), http.StatusInternalServerError)
+		http.Error(w, fmt.Sprintf(UnableToReadBodyErrFormat, err), http.StatusInternalServerError)
 		return
 	}
 
@@ -106,7 +73,7 @@ func (aH *APIHandler) saveSpan(w http.ResponseWriter, r *http.Request) {
 		// (NB): We decided to use this struct instead of straight batches to be as consistent with tchannel intake as possible.
 		batch := &tJaeger.Batch{}
 		if err = tdes.Read(batch, bodyBytes); err != nil {
-			http.Error(w, fmt.Sprintf(unableToReadBodyErrFormat, err), http.StatusBadRequest)
+			http.Error(w, fmt.Sprintf(UnableToReadBodyErrFormat, err), http.StatusBadRequest)
 			return
 		}
 		ctx, cancel := tchanThrift.NewContext(time.Minute)
@@ -117,8 +84,6 @@ func (aH *APIHandler) saveSpan(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-	case "zipkin.thrift":
-		handleZipkinThrift(aH.zipkinSpansHandler, bodyBytes, w)
 	default:
 		http.Error(w, fmt.Sprintf("Unsupported format type: %v", format), http.StatusBadRequest)
 		return
@@ -127,40 +92,3 @@ func (aH *APIHandler) saveSpan(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 }
 
-func handleZipkinThrift(zHandler ZipkinSpansHandler, bodyBytes []byte, w http.ResponseWriter) {
-	spans, err := deserializeZipkin(bodyBytes)
-	if err != nil {
-		http.Error(w, fmt.Sprintf(unableToReadBodyErrFormat, err), http.StatusBadRequest)
-		return
-	}
-
-	ctx, _ := tchanThrift.NewContext(time.Minute)
-	if _, err = zHandler.SubmitZipkinBatch(ctx, spans); err != nil {
-		http.Error(w, fmt.Sprintf("Cannot submit Zipkin batch: %v", err), http.StatusInternalServerError)
-		return
-	}
-}
-
-func deserializeZipkin(b []byte) ([]*zipkincore.Span, error) {
-	buffer := thrift.NewTMemoryBuffer()
-	buffer.Write(b)
-
-	transport := thrift.NewTBinaryProtocolTransport(buffer)
-	_, size, err := transport.ReadListBegin() // Ignore the returned element type
-	if err != nil {
-		return nil, err
-	}
-
-	// We don't depend on the size returned by ReadListBegin to preallocate the array because it
-	// sometimes returns a nil error on bad input and provides an unreasonably large int for size
-	var spans []*zipkincore.Span
-	for i := 0; i < size; i++ {
-		zs := &zipkincore.Span{}
-		if err = zs.Read(transport); err != nil {
-			return nil, err
-		}
-		spans = append(spans, zs)
-	}
-
-	return spans, nil
-}
