+++ cmd/standalone/main.go
@@ -44,6 +44,7 @@ import (
 	agentApp "github.com/uber/jaeger/cmd/agent/app"
 	basic "github.com/uber/jaeger/cmd/builder"
 	collectorApp "github.com/uber/jaeger/cmd/collector/app"
+	collectorZipkin "github.com/uber/jaeger/cmd/collector/app/zipkin"
 	collector "github.com/uber/jaeger/cmd/collector/app/builder"
 	queryApp "github.com/uber/jaeger/cmd/query/app"
 	query "github.com/uber/jaeger/cmd/query/app/builder"
@@ -132,12 +133,13 @@ func startCollector(logger *zap.Logger, baseFactory metrics.Factory, memoryStore
 	logger.Info("Starting jaeger-collector TChannel server", zap.Int("port", *collector.CollectorPort))
 
 	r := mux.NewRouter()
-	apiHandler := collectorApp.NewAPIHandler(jaegerBatchesHandler, zipkinSpansHandler)
-	apiHandler.RegisterRoutes(r)
 	httpPortStr := ":" + strconv.Itoa(*collector.CollectorHTTPPort)
 	recoveryHandler := recoveryhandler.NewRecoveryHandler(logger, true)
 
-	go zipkinCompEndpoint(logger, apiHandler, recoveryHandler)
+	go collectorZipkin.StartHttpAPI(logger, zipkinSpansHandler, recoveryHandler)
+
+	apiHandler := collectorApp.NewAPIHandler(jaegerBatchesHandler)
+	apiHandler.RegisterRoutes(r)
 
 	logger.Info("Starting jaeger-collector HTTP server", zap.Int("http-port", *collector.CollectorHTTPPort))
 	go func() {
@@ -145,16 +147,6 @@ func startCollector(logger *zap.Logger, baseFactory metrics.Factory, memoryStore
 			logger.Fatal("Could not launch jaeger-collector HTTP server", zap.Error(err))
 		}
 	}()
-}
-
-func zipkinCompEndpoint(logger *zap.Logger, apiHandler *collectorApp.APIHandler, recoveryHandler func (http.Handler) http.Handler) {
-	r := mux.NewRouter()
-	apiHandler.RegisterZipkinRoutes(r)
-	httpPortStr := ":" + strconv.Itoa(*collector.CollectorZipkinHTTPPort)
-	logger.Info("Listening for Zipkin HTTP traffic", zap.Int("zipkin.http-port", *collector.CollectorZipkinHTTPPort))
-	if err := http.ListenAndServe(httpPortStr, recoveryHandler(r)); err != nil {
-		logger.Fatal("Could not launch service", zap.Error(err))
-	}
 }
 
 func startQuery(logger *zap.Logger, baseFactory metrics.Factory, memoryStore *memory.Store) {
